requirements.txt
pyTelegramBotAPI
python-dotenv
openai
PyYAML
watchdog
flask